ENTITY FILES: Support multiple .ent files for each map.
Any .ent files in the maps or entfiles directories which are in format mapName#xxx.ent
will be added to the maps list.  If selected, sv_entityfile will be set to the name of
the entity file, and the map changed to mapName.

Requires server support for sv_entityfile cvar.