RACE: Allow race routes to be defined by map entities
This is to so KTX doesn't need to be updated whenever a new race map
is created.

Two new entity types:
  race_route_start
  race_route_marker

race_route_start requires the following properties:
  origin (vector)
  race_route_name (string)
  race_route_description (string)
  race_route_timeout (int)
  race_route_weapon_mode (enum - matches KTX raceWeapoMode_t)
  race_route_falsestart_mode (enum - matches KTX raceFalseStartMode_t)
  target (string, must point to corresponding race_route_marker)

race_route_marker requires:
  origin (vector)
  targetname (string)

Each entity links to next marker through target attribute, the last marker
in the sequence is considered the end of the route.