Make lower skilled bots easier
I'm not sure the justification for the lower bounds originally, but I
can't hang with even skill 0 bots.  I increased the bottom range of some
stuff that seemed related to accuracy by an order of magnitude.

With these changes bots in the low skill range miss A LOT more and
somewhere between level 3 and level 5 I go from able to keep up to
feeling quite challenged.