Improved map voting mechanics for admins.
A map vote from an admin will no longer bypass the percentage system
and force the map to change. It will be treated like a regular vote
instead.

If the need to force a change of map arises, there is a new command,
forcemap <mapname>, which does precisely that.

Only admins can execute forcemap, and it will refuse to work if a match
is currently in progress (admins can forcebreak the match first if
necessary).