TRAP: replace usage of visclient with trap_VisibleTo
This should help with FTE support.

As side note, visclient (and now trap_VisibleTo)
used in build in teamplay messages in KTX,
so during this task I realise what state of
teamplay messages in KTX - pretty BAD. They are broken
or incomplete or whatever. We either have to fix them
or remove. For example /tpmsg youtake fill in some buffer
but does not even call say.