RUNES: Let mappers dictate rune spawns.
More than once there has been discussions on rune spawns and how they
randomly might give one of the teams an upper hand.

This feature gives mappers the ability to integrate rune spawn positions
into the map design that will be used whenever a rune (re)spawns.

Dropping a rune still drops it on the ground, but if not picked up, the
regular timeout will cause the rune to respawn at a dedicated position if
such a position is available.